refactor: moving mocked context to __mocks__ folder
