+++ functions/__mocks__/index.ts
@@ -1,1 +1,2 @@
 export * from './mocked-context';
+export * from './mocked-event';
